requirements.txt
fastapi
uvicorn
websockets
google-genai
google-cloud-firestore